PPM files specified as P3/255 are integers
The Raytracing in One Weekend series uses integers in the PPM file, and
if we don't use integers as well some other programs such as ffmpeg
can not process the ppm files produced by this program

closes #1

+++ src/camera.rs
@@ -163,7 +163,9 @@ impl Camera {
                 ) * 256.;
                 format!(
                     "{} {} {}",
-                    color.x, color.y, color.z
+                    color.x as u8,
+                    color.y as u8,
+                    color.z as u8
                 )
             })
             .collect::<Vec<String>>()
